@@ -7,18 +7,20 @@
     "@material-ui/icons": "^3.0.1",
     "availity-reactstrap-validation": "^2.3.3",
     "axios": "^0.18.0",
-    "bootstrap": "^4.1.3",
+    "bootstrap": "^4.2.1",
     "geolib": "^2.0.24",
+    "install": "^0.12.2",
     "leaflet": "^1.3.4",
     "moment": "^2.24.0",
     "qrcode-reader": "^1.0.4",
-    "react": "^16.6.3",
+    "react": "^16.7.0",
     "react-bootstrap": "^0.32.4",
     "react-datepicker": "^2.0.0",
-    "react-dom": "^16.6.3",
+    "react-dom": "^16.7.0",
     "react-geolocation": "^1.0.4",
     "react-icons": "^3.3.0",
     "react-leaflet": "^2.1.2",
+    "react-loader-spinner": "^2.3.0",
     "react-moment": "^0.8.4",
     "react-qr-reader": "^2.1.2",
     "react-redux": "^5.1.1",
@@ -49,6 +51,5 @@
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ],
-  "proxy": "http://localhost:5000"
+  ]
 }
