@@ -1,8 +1,9 @@
 import React from 'react';
-import { Button, FormGroup, Label, Input, Breadcrumb, ListGroup } from 'reactstrap';
+import { Button, FormGroup, Label, Input, Breadcrumb, ListGroup, Container, Col, Row } from 'reactstrap';
 import { NavLink } from 'react-router-dom';
 const axios = require('axios');
 const moment = require('moment');
+
 
 export default class Classement extends React.Component {
   constructor(props) {
@@ -40,9 +41,9 @@ export default class Classement extends React.Component {
     return this.state.classement.map((classement, index) => {
       return (
         <div>
-          <Breadcrumb>
-            <ListGroup>
-              {classement.nom} : {classement.score} points ; Heure de fin : {classement.h_fin}
+          <Breadcrumb className="Breadcrumb">
+            <ListGroup className="ClassementText">
+              Équipe : {classement.nom} / Points : {classement.score} / Heure de fin : {classement.h_fin}
             </ListGroup>
           </Breadcrumb>
         </div>
@@ -53,18 +54,30 @@ export default class Classement extends React.Component {
   render() {
     return (
       <div>
-        <h1>Classement</h1>
-        <FormGroup>
-          <Label for="exampleSelect">Selectionnez le type de Classement</Label>
-          <Input type="select" name="tri" id="tri" onChange={this.saveChoice}>
-            <option value="indices" name="par temps">Choix...</option>
-            <option value="points" name="par temps">Par points</option>
-          </Input>
-        </FormGroup>
+        <Container>
+          <Row>
+            <Col className="text-left text-md-left">
+              <h1 className="titre">Classement des équipes</h1>
+            </Col>
+          </Row>
+          <Row>
+            <Col xs="0" md="4" />
+            <Col xs="12" md="5" className="text-left text-md-left">
+              <FormGroup>
+                <Label for="exampleSelect">Selectionnez le type de classement</Label>
 
-        {this.renderResult()}
+                <Input type="select" name="tri" id="tri" onChange={this.saveChoice}>
+                  <option value="indices" name="par temps">Choix...</option>
+                  <option value="points" name="par temps">Par points</option>
+                </Input>
+              </FormGroup>
+            </Col>
+          </Row>
 
-        <NavLink to={`/Admin/SessionPage/${window.localStorage.getItem('idAdmin')}`}><Button>Retour</Button></NavLink>
+          {this.renderResult()}
+
+          <NavLink to={`/Admin/SessionPage/${window.localStorage.getItem('idAdmin')}`}><Button className="text-center text-md-center">Retour</Button></NavLink>
+        </Container>
       </div>
     );
   }
