@@ -1,6 +1,6 @@
 import React from 'react'
 import axios from 'axios'
-import { InputGroup, InputGroupAddon, Input, Button } from 'reactstrap'
+import { InputGroup, InputGroupAddon, Input, Button, FormGroup, Label, Container, Row, Col } from 'reactstrap'
 import logo from './logo_tinyplanet_orange.png'
 import './Connexion.css'
 import { NavLink } from 'react-router-dom'
@@ -44,7 +44,7 @@ export default class AdminComptes extends React.Component {
                 if (response.data.success === true) {
                     window.localStorage.setItem("tokenAdmin", response.data.token)
                     window.localStorage.setItem("idAdmin", response.data.id)
-                    window.location.href =`/Admin/SessionPage/${response.data.id}`
+                    window.location.href = `/Admin/SessionPage/${response.data.id}`
                 } else alert("Formulaire érroné")
             })
             .catch(function (error) {
@@ -55,17 +55,26 @@ export default class AdminComptes extends React.Component {
     render() {
         return (
             <div className='containerConnexion'>
+                <Container>
+                    <Row>
+                        <Col xs="0" md="2" />
+                        <Col xs="12" md="8">
+                            <img src={logo} />
 
-                <img src={logo} />
-
-                <InputGroup >
-                    <InputGroupAddon addonType="prepend" >Adresse email :</InputGroupAddon>
-                    <Input onChange={this.modifyEmail} />
-                    <InputGroupAddon addonType="prepend" >Mot de passe :</InputGroupAddon>
-                    <Input onChange={this.modifyPassword} />
-                </InputGroup>
-
-                <Button color="secondary" size="lg" onClick={this.submitLogin}>Connexion</Button>
+                            <InputGroup >
+                                <FormGroup>
+                                    <Label for="exampleEmail">Email</Label>
+                                    <Input onChange={this.modifyEmail} type="email" name="email" id="exampleEmail" />
+                                </FormGroup>
+                                <FormGroup>
+                                    <Label for="examplePassword">Mot de passe</Label>
+                                    <Input onChange={this.modifyPassword} type="password" name="password" id="examplePassword"  />
+                                </FormGroup>
+                            </InputGroup>
+                            <Button color="secondary" className="text-center text-md-center" size="lg" onClick={this.submitLogin}>Connexion</Button>
+                        </Col>
+                    </Row>
+                </Container>
             </div >
         );
     }
