@@ -1,20 +1,24 @@
 import React from 'react'
 import axios from 'axios'
 import Toggle from "react-toggle-component"
-import { Breadcrumb, BreadcrumbItem } from 'reactstrap'
+import { Button } from 'reactstrap'
+import logo from './logo_tinyplanet_orange.png'
 import './Connexion.css'
+import './SessionPage.css'
+import { FaLock } from 'react-icons/fa'
 
 
 export default class AdminComptes extends React.Component {
     constructor(props) {
         super(props)
         this.state = {
-            checked: null,
-            etat: null,
+            checked: true,
+            etat: "désactivée"
         }
     }
 
     componentDidMount() {
+        //this.modifyActivation()
         axios.get('http://localhost:5000/api/session')
             .then(response => {
                 console.log(response.data[0])
@@ -31,12 +35,12 @@ export default class AdminComptes extends React.Component {
         
     }
 
-    async modifyActivation(value) {
-        await this.setState({
-            checked: value
+    modifyActivation = () => {
+        this.setState({
+            checked: !this.state.checked
         })
-        await this.state.checked ? this.setState({ etat: "activée" }) : this.setState({ etat: "désactivée" })
-        await axios.put(`http://localhost:5000/api/session/activation`,
+        this.state.checked ? this.setState({ etat: "activée" }) : this.setState({ etat: "désactivée" })
+        axios.put(`http://localhost:5000/api/session/activation`,
             {
                 isactivate: this.state.checked
             })
@@ -50,16 +54,14 @@ export default class AdminComptes extends React.Component {
 
     render() {
         return (
-
-            <Breadcrumb>
-                <BreadcrumbItem >Session {this.state.etat}</BreadcrumbItem>
-                <Toggle
-                    name="activation"
-                    mode="switch"
-                    checked={this.state.checked}
-                    onToggle={value => this.modifyActivation(value)} />
-            </Breadcrumb>
-
-        );
+            <div className="toggleBlock">
+                <Button className={this.state.etat === "activée" ? "toggleActived" : "toggleDeactivated"} onClick={this.modifyActivation} >
+                    <div className='textButtonIcon'>
+                        <FaLock />
+                    </div>
+                    <div className="togglerButtonText">{this.state.etat}</div>
+                </Button>
+            </div>
+        )
     }
 }
