@@ -3,6 +3,10 @@ import { Button, Alert, Input, FormGroup, Label, FormText } from 'reactstrap';
 import Editable from 'react-x-editable';
 import { NavLink } from 'react-router-dom';
 import './UnEnigme.css';
+import './ListEnigmes.css';
+import './SessionPage.css';
+import "react-toggle-component/styles.css";
+import { FaScroll, FaChevronLeft } from 'react-icons/fa'
 const axios = require('axios');
 
 
@@ -160,16 +164,23 @@ export default class UnEnigme extends Component {
 
     render() {
         return (
-            <div>
+            <Button>
+            <div className='textButtonUne'>
 
                 {this.state.titre ?
                     <div>
-                        <Alert color="dark">
-                            Id de l'énigme : {this.state.id}
+                        <Alert 
+                        color="dark">
+                            Id Enigme {this.state.id}
                         </Alert>
-                        <Alert color="dark">
-                            Titre :
+
+                        <Alert 
+                        className="id"
+                        color="dark">
+                        
+                        <p className="titre">Titre : </p>
                             <Editable
+                                className="id"
                                 name="username"
                                 dataType="text"
                                 value={this.state.titre}
@@ -185,9 +196,11 @@ export default class UnEnigme extends Component {
                             />
                         </Alert>
 
-                        <Alert color="dark">
-                            Points à gagner :
+                        <Alert className="id"
+                        color="dark">
+                            <p className="titre">Points à gagner :</p>
                             <Editable
+                                className="titre"
                                 name="agagner"
                                 dataType="number"
                                 value={this.state.agagner}
@@ -203,24 +216,10 @@ export default class UnEnigme extends Component {
                             />
                         </Alert>
 
-                        <Alert color="dark">
-                            Image utilisée (Ne fonctionne pas pour l'instant): <Editable
-                                name="username"
-                                dataType="text"
-                                value={this.state.img}
-                                validate={(value) => {
-                                    if (!value) {
-                                        return 'Required';
-                                    }
-                                    else {
-                                        this.modifyImage(value)
-                                    }
-                                }
-                                }
-                            />
-                        </Alert>
-                        <Alert color="dark">
-                            Énoncé : <Editable
+                        
+                        <Alert className="id"
+                        color="dark">
+                             <p className="titre">Énoncé :</p><Editable
                                 name="username"
                                 dataType="textarea"
                                 value={this.state.enonce}
@@ -236,8 +235,9 @@ export default class UnEnigme extends Component {
                             />
 
                         </Alert>
-                        <Alert color="dark">
-                            Question : <Editable
+                        <Alert className="id"
+                        color="dark">
+                             <p className="titre">Question : </p><Editable
                                 name="username"
                                 dataType="textarea"
                                 value={this.state.question}
@@ -252,8 +252,9 @@ export default class UnEnigme extends Component {
                                 }
                             />
                         </Alert>
-                        <Alert color="dark">
-                            Indices :
+                        <Alert className="id"
+                        color="dark">
+                            <p className="titre"> Indices : </p>
                         <Editable
                                 name="username"
                                 dataType="text"
@@ -297,8 +298,9 @@ export default class UnEnigme extends Component {
                                 }
                             />
                         </Alert>
-                        <Alert color="dark">
-                            Réponse :
+                        <Alert className="id"
+                        color="dark">
+                            <p className="titre"> Réponse : </p>
                             <Editable
                                 name="username"
                                 dataType="text"
@@ -314,9 +316,10 @@ export default class UnEnigme extends Component {
                                 }
                             />
                         </Alert>
-                        <h3>Informations géographiques du lieu :</h3>
-                        <Alert color="dark">
-                            Lattitude : <Editable
+                        <h3 className="edit">Informations géographiques du lieu :</h3>
+                        <Alert className="id"
+                        color="dark">
+                            <p className="titre">Lattitude : </p> <Editable
                                 name="username"
                                 dataType="text"
                                 value={this.state.coordonnees[0]}
@@ -330,7 +333,7 @@ export default class UnEnigme extends Component {
                                 }
                                 }
                             />
-                            Longitude :
+                            <p className="titre">Longitude : </p>
                             <Editable
                                 name="username"
                                 dataType="text"
@@ -346,8 +349,9 @@ export default class UnEnigme extends Component {
                                 }
                             />
                         </Alert>
-                        <Alert color="dark">
-                            Précautions sur le lieu :
+                        <Alert className="id"
+                        color="dark">
+                            <p className="titre">Précautions sur le lieu : </p>
                             <Editable
                                 name="username"
                                 dataType="text"
@@ -366,10 +370,16 @@ export default class UnEnigme extends Component {
                     </div>
                     : null}
                 <Button className={this.state.button} onClick={this.sendModifications}>Valider les modifications</Button>
+                < Button className="buttonMenu">
                 <NavLink to={`/Admin/ListEnigmes/${window.localStorage.getItem('idAdmin')}`}>
-                    <Button>Retour</Button>
+                <p className="textButton">Retour</p>
                 </NavLink>
+                <div className='textButtonIcon'>
+                        <FaChevronLeft />
+                    </div>
+                </Button>
             </div>
+            </Button>
         );
     }
 }
